split python scripts into seperate files
