+++ airflow/standalone_scripts/af2_pickle_fix_plugin.py
@@ -1,0 +1,6 @@
+from airflow.executors.local_executor import LocalWorker, QueuedLocalWorker
+
+QueuedLocalWorker.do_work.__name__ = "do_work"
+QueuedLocalWorker.do_work.__qualname__ = "QueuedLocalWorker.do_work"
+LocalWorker.do_work.__name__ = "do_work"
+LocalWorker.do_work.__qualname__ = "LocalWorker.do_work"
